+++ Main/Controllers/InventoryControllers/ControllerInterface/IIngredientInventoryController.cs
@@ -10,9 +10,9 @@ namespace Main.Controllers.InventoryControllers.ControllerInterface
 {
     public interface IIngredientInventoryController
     {
-        EntityResult<string> Delete(int ingredeintId, long inventoryId, string remarks);
+        EntityResult<string> Delete(long ingredeintId, long inventoryId, string remarks);
         EntityResult<IngredientInventoryModel> Save(IngredientInventoryModel input, bool isNew, string remarks);
-        EntityResult<IngredientInventoryModel> IncreaseQtyValue(int ingredientId, long inventoryId, decimal qtyValue, string remarks);
-        EntityResult<IngredientInventoryModel> DecreaseQtyValue(int ingredientId, long inventoryId, decimal qtyValue, string remarks);
+        EntityResult<IngredientInventoryModel> IncreaseQtyValue(long ingredientId, long inventoryId, decimal qtyValue, string remarks);
+        EntityResult<IngredientInventoryModel> DecreaseQtyValue(long ingredientId, long inventoryId, decimal qtyValue, string remarks);
     }
 }
